+++ docs/project/changelog.md
@@ -42,6 +42,8 @@ substitutions:
 - {{ Fix }} The build will error out earlier if `cmake` or `libtool` are not installed.
   {pr}`2423`
 
+- {{ Feature }} Added `pyodide.run_js` API.
+  {pr}`2426`
 - {{ Enhancement }} Add SHA-256 hash of package to entries in `packages.json`
   {pr}`2455`